Changed toggled to whileHeld

+++ ERNIE/src/org/usfirst/frc/team4750/robot/RobotMap.java
@@ -19,7 +19,7 @@ public class RobotMap {
 	public static final int SHOOTER_MOTOR = 7; //subject to change
 	
 	//Motor Speeds
-	public static final int SHOOTER_MOTOR_SPEED = 1;
+	public static final double SHOOTER_MOTOR_SPEED = 1;
 	
 	// For example to map the left and right motors, you could define the
 	// following variables to use with your drivetrain subsystem.
